+++ src/Operations/Decorators/ConsulterWithDanfse.php
@@ -40,7 +40,7 @@ final readonly class ConsulterWithDanfse implements ConsultsNfse
 
     public function eventos(
         string $chave,
-        TipoEvento|int $tipoEvento = TipoEvento::CancelamentoPorIniciativaPrestador,
+        TipoEvento|int $tipoEvento = TipoEvento::Cancelamento,
         int $nSequencial = 1,
     ): EventsResponse {
         return $this->inner->eventos($chave, $tipoEvento, $nSequencial);